# backend/internal/httpapi/openapi.yaml
openapi: 3.1.0
info:
  title: Home Finance API
  version: "1.4.0"
  description: |
    Internal API for a personal/family budgeting app. Plain net/http + chi
    router, hand-written JSON handlers. All money amounts are integer cents.
    Dates are YYYY-MM-DD strings; timestamps are RFC3339.

    Every error response uses the shape
    `{"error": {"code": "...", "message": "..."}}`. Messages are safe to show
    to a user; `db_error` carries a fixed generic message and never raw driver
    text.

    JSON request bodies are capped at 1 MiB by middleware; a larger body gets
    413 `body_too_large`. Multipart uploads bypass that cap and enforce their
    own, larger limit.

    Unmatched `/api` paths return 404 `not_found` and unsupported methods
    return 405 `method_not_allowed`, both as JSON rather than the SPA.

    Operations marked "Admin only" also require the caller's role to be
    `admin`; other roles get 403 `forbidden`.
servers:
  - url: /
...
security:
  - cookieAuth: []
    csrfHeader: []
paths:
...
  /auth/login:
    get:
      operationId: authLogin
      tags: [Auth]
      summary: Redirects to the OIDC provider (or fakes dev login) and sets session state.
      security: []
      parameters:
        - name: return_to
          in: query
          description: Path to redirect back to after auth completes. Sanitized to same-origin relative paths only.
          schema:
            type: string
      responses:
        "302":
          description: Redirect to OIDC provider (prod) or return_to (dev fake auth).
  /auth/callback:
    get:
      operationId: authCallback
      tags: [Auth]
      summary: OIDC callback. Validates state/nonce, exchanges code, upserts user, establishes session. Also completes the /auth/reauth step-up flow (distinguished internally by which session-stored state the callback's state param matches) — on that path it verifies the returned identity matches the current session before granting a fresh reauthAt, instead of establishing a new session.
      security: []
      parameters:
        - name: state
          in: query
          schema:
            type: string
        - name: code
          in: query
          schema:
            type: string
      responses:
        "302":
          description: Redirect to sanitized return_to path on success.
        "400":
          description: Invalid state or nonce.
        "401":
          description: Code exchange failed.
        "403":
          description: Email not in allowed list.
        "500":
          description: Claims decode or DB upsert error.
  /auth/reauth:
    get:
      operationId: authReauth
      tags: [Auth]
      summary: Starts a step-up re-authentication popup (prompt=login) used to confirm identity before a destructive action. Requires an existing session and a same-origin (or absent) Sec-Fetch-Site. Rate-limited per session user, falling back to remote address.
      responses:
        "302":
          description: Redirect to OIDC provider with prompt=login (prod), or an immediate self-closing confirmation page (dev fake auth).
        "401":
          description: Not logged in.
        "403":
          description: Sec-Fetch-Site indicates a cross-site navigation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/logout:
    post:
      operationId: authLogout
      tags: [Auth]
      summary: Destroys the session and redirects to /login. Requires the X-Requested-With CSRF header.
      security:
        - csrfHeader: []
      responses:
        "302":
          description: Redirect to /login.
        "403":
          description: Missing X-Requested-With header.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/me:
    get:
      operationId: getMe
      tags: [Auth]
      summary: Returns the current authenticated user.
      responses:
        "200":
          description: Current user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      operationId: updateMe
      tags: [Auth]
      summary: Updates the caller's own display name.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMeRequest"
      responses:
        "204":
          description: Display name updated.
        "400":
          description: Empty displayName, or longer than 200 characters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: User not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          $ref: "#/components/responses/BodyTooLarge"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/reauth-status:
    get:
      operationId: getReauthStatus
      tags: [Auth]
      summary: Polled by the frontend while the /auth/reauth popup is open, to detect when step-up reauth completes and auto-close it.
      responses:
        "200":
          description: Freshness of the current step-up reauth.
          content:
            application/json:
              schema:
                type: object
                properties:
                  fresh:
                    type: boolean
                required: [fresh]
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/users:
    get:
      operationId: listUsers
      tags: [Users]
      summary: Lists all users with their role. Admin only.
      responses:
        "200":
          description: Users.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/users/{id}/role:
    patch:
      operationId: updateUserRole
      tags: [Users]
      summary: Sets a user's role. Admin only. The last remaining admin cannot be demoted.
      parameters:
        - $ref: "#/components/parameters/EntityId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRoleRequest"
      responses:
        "204":
          description: Role updated (no-op when already set).
        "400":
          description: Invalid id, or role is neither "admin" nor "user".
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: User not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: "`last_admin`: cannot demote the last remaining admin."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          $ref: "#/components/responses/BodyTooLarge"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/periods:
    get:
      operationId: listPeriods
      tags: [Periods]
      summary: Lists periods for a given year with income/expense/surplus totals.
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Periods for the year.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Period"
        "400":
          description: Missing or invalid year.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createPeriod
      tags: [Periods]
      summary: Creates a period for year/month, optionally copying budget lines/splits/incomes from a prior period.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePeriodRequest"
      responses:
        "201":
          description: Period created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PeriodBasic"
        "400":
          description: Invalid body, or year/month out of range.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Year is locked, or period already exists for that month.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          $ref: "#/components/responses/BodyTooLarge"
        "500":
          description: Server/DB error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/periods/{id}/overview:
    get:
      operationId: getPeriodOverview
      tags: [Periods]
      summary: Full period detail - incomes, budget lines (with effective/transaction totals), pot splits, and totals.
      parameters:
        - $ref: "#/components/parameters/PeriodId"
      responses:
        "200":
          description: Period overview.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PeriodOverview"
        "400":
          description: Invalid id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Period not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/periods/{id}/close:
    post:
      operationId: closePeriod
      tags: [Periods]
      summary: Closes a period, allocating surplus to pots and generating carryover ledger/income entries for next month. Admin only.
      parameters:
        - $ref: "#/components/parameters/PeriodId"
      responses:
        "204":
          description: Period closed.
        "400":
          description: Invalid id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Period not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Period already closed, year locked, next period already closed, or the next year is locked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server/DB error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/periods/{id}/reopen:
    post:
      operationId: reopenPeriod
      tags: [Periods]
      summary: Reopens a closed period, removing generated carryover ledger entries and carryover income entries. Admin only.
      parameters:
        - $ref: "#/components/parameters/PeriodId"
      responses:
        "204":
          description: Period reopened.
        "400":
          description: Invalid id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Period not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Period already open, year is locked, or next period is closed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server/DB error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/periods/{id}:
    delete:
      operationId: deletePeriod
      tags: [Periods]
      summary: Deletes an open period after a fresh Authentik re-authentication. Admin only.
      parameters:
        - $ref: "#/components/parameters/PeriodId"
      responses:
        "204":
          description: Period deleted.
        "400":
          description: Invalid id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: No fresh re-authentication on file; client should trigger the /auth/reauth popup and retry.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Period not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Period is closed, year is locked, or the previous period's carryover lands in it (`carryover_present`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server/DB error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/audit-log:
    get:
      operationId: listAuditLog
      tags: [AuditLog]
      summary: Lists audit log entries, newest first. Admin only.
      parameters:
        - name: limit
          in: query
          description: Page size, 1-200. Out-of-range or unparseable values fall back to 50.
          schema:
            type: integer
            default: 50
        - name: before
          in: query
          description: Return only entries with a lower id (keyset pagination).
          schema:
            type: integer
            format: int64
        - name: action
          in: query
          schema:
            type: string
        - name: entityType
          in: query
          schema:
            type: string
        - name: userEmail
          in: query
          schema:
            type: string
        - name: from
          in: query
          description: Inclusive lower bound on createdAt.
          schema:
            type: string
            format: date
        - name: to
          in: query
          description: Inclusive upper bound on createdAt.
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Audit log entries.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLogEntry"
        "400":
          description: Malformed from/to date; expected YYYY-MM-DD.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          description: Server/DB error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/periods/{id}/incomes:
    post:
      operationId: createIncomeEntry
      tags: [Incomes]
      summary: Adds a normal income entry to a period. Fails if the period is closed or its year is locked.
      parameters:
        - $ref: "#/components/parameters/PeriodId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIncomeEntryRequest"
      responses:
        "201":
          description: Income entry created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeEntry"
        "400":
          description: Invalid id, body, or amount.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Period not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Period closed or year locked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          $ref: "#/components/responses/BodyTooLarge"
        "500":
          description: Server/DB error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/incomes/{id}:
    put:
      operationId: updateIncomeEntry
      tags: [Incomes]
      summary: Updates an income entry's label/amount/notes/sort order.
      parameters:
        - $ref: "#/components/parameters/EntityId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIncomeEntryRequest"
      responses:
        "204":
          description: Income entry updated.
        "400":
          description: Invalid id, body, or amount.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Income entry not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Period closed or year locked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          $ref: "#/components/responses/BodyTooLarge"
        "500":
          description: Server/DB error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteIncomeEntry
      tags: [Incomes]
      summary: Deletes an income entry.
      parameters:
        - $ref: "#/components/parameters/EntityId"
      responses:
        "204":
          description: Income entry deleted.
        "400":
          description: Invalid id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Income entry not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Period closed, year locked, or the entry has a non-zero amount or transactions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/periods/{id}/budget-lines:
    post:
      operationId: createBudgetLine
      tags: [BudgetLines]
      summary: Adds a budget line to a period. Fails if the period is closed or its year is locked.
      parameters:
        - $ref: "#/components/parameters/PeriodId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBudgetLineRequest"
      responses:
        "201":
          description: Budget line created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetLine"
        "400":
          description: Invalid id, body, amount, or a child category used as the line's category.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Period not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Period closed or year locked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          $ref: "#/components/responses/BodyTooLarge"
        "500":
          description: Server/DB error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/budget-lines/{id}:
    put:
      operationId: updateBudgetLine
      tags: [BudgetLines]
      summary: Updates a budget line's label/amount/tracking flag/sort order.
      parameters:
        - $ref: "#/components/parameters/EntityId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBudgetLineRequest"
      responses:
        "204":
          description: Budget line updated.
        "400":
          description: Invalid id, body, or amount.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Budget line not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Period closed or year locked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          $ref: "#/components/responses/BodyTooLarge"
        "500":
          description: Server/DB error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteBudgetLine
      tags: [BudgetLines]
      summary: Deletes a budget line.
      parameters:
        - $ref: "#/components/parameters/EntityId"
      responses:
        "204":
          description: Budget line deleted.
        "400":
          description: Invalid id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Budget line not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Period closed, year locked, or the line has a non-zero amount or transactions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/periods/{id}/transactions:
    get:
      operationId: listTransactions
      tags: [Transactions]
      summary: Lists transactions for a period, optionally filtered by category.
      parameters:
        - $ref: "#/components/parameters/PeriodId"
        - name: categoryId
          in: query
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Transactions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
        "400":
          description: Invalid id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createTransaction
      tags: [Transactions]
      summary: Adds a transaction to a period. Fails if the period is closed or its year is locked.
      parameters:
        - $ref: "#/components/parameters/PeriodId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionRequest"
      responses:
        "201":
          description: Transaction created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Invalid id, body, or amount.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Period not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Period closed or year locked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          $ref: "#/components/responses/BodyTooLarge"
        "500":
          description: Server/DB error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/category-description-presets/{id}:
    delete:
      operationId: deleteCategoryDescriptionPreset
      tags: [Categories]
      summary: Deletes a category description preset. Admin only.
      parameters:
        - $ref: "#/components/parameters/EntityId"
      responses:
        "204":
          description: Preset deleted.
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Preset not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Session cookie set after successful login (OIDC or dev fake auth).
    csrfHeader:
      type: apiKey
      in: header
      name: X-Requested-With
      description: Required (value "XMLHttpRequest") on all mutating /api requests alongside the session cookie.
  parameters:
...
    EntityId:
      name: id
      in: path
      required: true
      description: Resource id.
      schema:
        type: integer
        format: int64
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [error]
...